compiler/checker: move 'show' expression checking out of the compiler
This change moves the 'show' expression checkout from the compiler to
the templates package.
