+++ compiler/checker_util.go
@@ -11,7 +11,6 @@ import (
 	"fmt"
 	"reflect"
 	"strconv"
-	"time"
 	"unicode"
 	"unicode/utf8"
 
@@ -140,44 +139,6 @@ var operatorsOfKind = [...][22]bool{
 	reflect.String:     stringOperators,
 	reflect.Interface:  interfaceOperators,
 }
-
-type (
-	HTMLStringer interface{ HTML() string }
-	CSSStringer  interface{ CSS() string }
-	JSStringer   interface{ JS() string }
-	JSONStringer interface{ JSON() string }
-)
-
-// These interfaces are like HTMLStringer, CSSStringer, JSStringer and
-// JSONStringer, but their method accepts a runtime.Env parameter that can be used inside the
-// method's body to access some environment information.
-type (
-	HTMLEnvStringer interface{ HTML(runtime.Env) string }
-	CSSEnvStringer  interface{ CSS(runtime.Env) string }
-	JSEnvStringer   interface{ JS(runtime.Env) string }
-	JSONEnvStringer interface{ JSON(runtime.Env) string }
-)
-
-// EnvStringer is like fmt.Stringer, but its method accepts a runtime.Env as
-// parameter.
-type EnvStringer interface{ String(runtime.Env) string }
-
-var stringerType = reflect.TypeOf((*fmt.Stringer)(nil)).Elem()
-var envStringerType = reflect.TypeOf((*EnvStringer)(nil)).Elem()
-
-var htmlStringerType = reflect.TypeOf((*HTMLStringer)(nil)).Elem()
-var htmlEnvStringerType = reflect.TypeOf((*HTMLEnvStringer)(nil)).Elem()
-
-var cssStringerType = reflect.TypeOf((*CSSStringer)(nil)).Elem()
-var cssEnvStringerType = reflect.TypeOf((*CSSEnvStringer)(nil)).Elem()
-
-var jsStringerType = reflect.TypeOf((*JSStringer)(nil)).Elem()
-var jsEnvStringerType = reflect.TypeOf((*JSEnvStringer)(nil)).Elem()
-
-var jsonStringerType = reflect.TypeOf((*JSONStringer)(nil)).Elem()
-var jsonEnvStringerType = reflect.TypeOf((*JSONEnvStringer)(nil)).Elem()
-
-var timeType = reflect.TypeOf(time.Time{})
 
 // convert implicitly converts an untyped value. If the converted value is a
 // constant, convert returns its value, otherwise returns nil.
@@ -678,112 +639,6 @@ func operatorFromAssignmentType(assignmentType ast.AssignmentType) ast.OperatorT
 		return ast.OperatorRightShift
 	}
 	panic("unexpected assignment type")
-}
-
-// printedAsJS reports whether a type can be printed as JavaScript. It returns
-// an error if the type cannot be printed.
-func printedAsJS(t reflect.Type) error {
-	kind := t.Kind()
-	if reflect.Bool <= kind && kind <= reflect.Float64 || kind == reflect.String ||
-		t == timeType ||
-		t.Implements(jsStringerType) ||
-		t.Implements(jsEnvStringerType) ||
-		t.Implements(errorType) {
-		return nil
-	}
-	switch kind {
-	case reflect.Array:
-		if err := printedAsJS(t.Elem()); err != nil {
-			return fmt.Errorf("array of %s cannot be printed as JavaScript", t.Elem())
-		}
-	case reflect.Interface:
-	case reflect.Map:
-		key := t.Key().Kind()
-		switch {
-		case key == reflect.String:
-		case reflect.Bool <= key && key <= reflect.Complex128:
-		case t.Implements(stringerType):
-		case t.Implements(envStringerType):
-		default:
-			return fmt.Errorf("map with %s key cannot be printed as JavaScript", t.Key())
-		}
-		err := printedAsJS(t.Elem())
-		if err != nil {
-			return fmt.Errorf("map with %s element cannot be printed as JavaScript", t.Elem())
-		}
-	case reflect.Ptr, reflect.UnsafePointer:
-		return printedAsJS(t.Elem())
-	case reflect.Slice:
-		if err := printedAsJS(t.Elem()); err != nil {
-			return fmt.Errorf("slice of %s cannot be printed as JavaScript", t.Elem())
-		}
-	case reflect.Struct:
-		n := t.NumField()
-		for i := 0; i < n; i++ {
-			field := t.Field(i)
-			if field.PkgPath == "" {
-				if err := printedAsJS(field.Type); err != nil {
-					return fmt.Errorf("struct containing %s cannot be printed as JavaScript", field.Type)
-				}
-			}
-		}
-	default:
-		return fmt.Errorf("type %s cannot be printed as JavaScript", t)
-	}
-	return nil
-}
-
-// printedAsJSON reports whether a type can be printed as JSON. It returns an
-// error if the type cannot be printed.
-func printedAsJSON(t reflect.Type) error {
-	kind := t.Kind()
-	if reflect.Bool <= kind && kind <= reflect.Float64 || kind == reflect.String ||
-		t == timeType ||
-		t.Implements(jsonStringerType) ||
-		t.Implements(jsonEnvStringerType) ||
-		t.Implements(errorType) {
-		return nil
-	}
-	switch kind {
-	case reflect.Array:
-		if err := printedAsJSON(t.Elem()); err != nil {
-			return fmt.Errorf("array of %s cannot be printed as JSON", t.Elem())
-		}
-	case reflect.Interface:
-	case reflect.Map:
-		key := t.Key().Kind()
-		switch {
-		case key == reflect.String:
-		case reflect.Bool <= key && key <= reflect.Complex128:
-		case t.Implements(stringerType):
-		case t.Implements(envStringerType):
-		default:
-			return fmt.Errorf("map with %s key cannot be printed as JSON", t.Key())
-		}
-		err := printedAsJSON(t.Elem())
-		if err != nil {
-			return fmt.Errorf("map with %s element cannot be printed as JSON", t.Elem())
-		}
-	case reflect.Ptr, reflect.UnsafePointer:
-		return printedAsJSON(t.Elem())
-	case reflect.Slice:
-		if err := printedAsJSON(t.Elem()); err != nil {
-			return fmt.Errorf("slice of %s cannot be printed as JSON", t.Elem())
-		}
-	case reflect.Struct:
-		n := t.NumField()
-		for i := 0; i < n; i++ {
-			field := t.Field(i)
-			if field.PkgPath == "" {
-				if err := printedAsJSON(field.Type); err != nil {
-					return fmt.Errorf("struct containing %s cannot be printed as JSON", field.Type)
-				}
-			}
-		}
-	default:
-		return fmt.Errorf("type %s cannot be printed as JSON", t)
-	}
-	return nil
 }
 
 // removeEnvArg returns a type equal to typ but with the vm environment
